version: remove BuildTime
`BuildTime` variable, related function and tests have been removed.

Adding build time to the binary is a source of build reproducibility
issue.
Given that the resulting binary contains the date and time of it's build
from the build server (or developer laptop) reproducing it is very
difficult.

Using the source date approach proposed by
https://reproducible-builds.org/ allows to overcome this limitation
while retaining an interesting element of versioning (the date of the
code included in the binary).

For further details see:
https://reproducible-builds.org/
https://reproducible-builds.org/docs/timestamps/
https://reproducible-builds.org/docs/source-date-epoch/