fragment: timestamp function should return different values
The previous default value of `timestamp` function (`time.Now().Unix`)
always returned the same value, taken at the moment of `Creator`
initialization.

This was not the intended behaviour as the `Creator.Create()` method
supports a `slug` argument, making it possible to create multiple
fragments using a single `Creator`.

To solve this the `timestamp()` function now returns a `time.Time`, that
can be converted to a UNIX timestamp when needed (when calling
`filename()`).

Tests are provided to ensure multiple calls to `timestamp()` produce a
different result.