github: vcr testing http client
go-vcr is a recorder/replayer library for testing libraries that perform
HTTP requests. Instead of mocking the only change to be performed is
replacing the transport with the recorder in the tests, something that
go-github library supports.

Calls are recorded in YAML files in `testdata/fixtures` (named with the
test name) and used to reply the HTTP response. This effectively stop
the tests from calling external APIs while reducing to the minimum the
impact in the code.

Once a fixture exists, is not overwritten. To update a fixture delete
the corresponding file and run tests again.