provide shared utility for viper<>cobra
There is no proper way to bind a cobra flag with viper config
value.

Of the multiple possible ways the one in this PR is the less
unexpected one, provided some context around how it works and
why is needed.

Using `BindPFlag` does not work because when binding a flag with
the same name in multiple commands only the last binding would be
respected; this effectively prevents the propagation of the flag
value to viper in such cases and is very hard to debug (code works
before and breaks after an unrelated code change).